backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
requests
openai
python-multipart

